Пофиксил выбор даты и удалил лишние формы
